+++ src/sass/functions/map.rs
@@ -96,7 +96,11 @@ pub fn create_module() -> Scope {
     });
     def!(f, keys(map), |s| {
         let map = get_map(s, name!(map))?;
-        Ok(Value::List(map.keys(), Some(ListSeparator::Comma), false))
+        Ok(Value::List(
+            map.keys().cloned().collect(),
+            Some(ListSeparator::Comma),
+            false,
+        ))
     });
     def_va!(f, merge(map1, map2), |s| {
         let mut map1 = get_map(s, name!(map1))?;
@@ -145,7 +149,11 @@ pub fn create_module() -> Scope {
     def_va!(f, set(map, args), set);
     def!(f, values(map), |s| {
         let map = get_map(s, name!(map))?;
-        Ok(Value::List(map.values(), Some(ListSeparator::Comma), false))
+        Ok(Value::List(
+            map.values().cloned().collect(),
+            Some(ListSeparator::Comma),
+            false,
+        ))
     });
     f
 }